visual changes made and delete game mutation fixed

+++ src/filters/date.js
@@ -1,4 +1,4 @@
 export default (value) => {
   const date = new Date(value)
-  return date.toLocaleString('pl-PL', {weekday: 'long', month: 'long', day: 'numeric', year: 'numeric'})
+  return date.toLocaleString('pl-PL', {weekday: 'long', month: 'numeric', day: 'numeric', year: 'numeric'})
 }
